[UI] Phase 4: Consolidate rjsf-utils and typing-filter into shared/FormFields (#18659)
Move two small legacy folders under `ui/components/` into the shared
form-field area so RJSF-adjacent code lives together:

- `ui/components/rjsf-utils/common.ts`
  -> `ui/components/shared/FormFields/rjsf-utils/common.ts`
- `ui/components/typing-filter/**`
  -> `ui/components/shared/FormFields/typing-filter/**`

All moves use `git mv` to preserve history. Importers are rewritten to
the `@/components/shared/FormFields/...` alias form. Two eslint legacy
offender allowlist entries are updated to the new paths.

Drops the top-level `ui/components/` folder count by 2 (23 -> 21,
target ~15).

Refs #18659
Refs #18654
