+++ ui/eslint.config.js
@@ -106,7 +106,7 @@ const legacyLiteralColorOffenders = [
   'components/registry/helper.ts',
   'components/workspaces/SpacesSwitcher/WorkspaceSwitcher.tsx',
   'components/workspaces/SpacesSwitcher/styles.tsx',
-  'components/typing-filter/style.tsx',
+  'components/shared/FormFields/typing-filter/style.tsx',
   'components/user-preferences/index.tsx',
   'components/user-preferences/style.tsx',
   'components/designs/configurator/CustomBreadCrumb.tsx',
@@ -296,7 +296,7 @@ const legacyInlineStyleOffenders = [
   'components/workspaces/SpacesSwitcher/WorkspaceSwitcher.tsx',
   'components/workspaces/SpacesSwitcher/components.tsx',
   'components/TroubleshootingComponent.tsx',
-  'components/typing-filter/index.tsx',
+  'components/shared/FormFields/typing-filter/index.tsx',
   'components/user-preferences/index.tsx',
   'components/shared/Modal/ViewInfoModal.tsx',
   'components/ViewSwitch.tsx',
